Organized files into respective directories.
