+++ src/components/MoviePoster/MoviePoster.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { withRouter, NavLink } from 'react-router-dom';
-import { loadFavorites } from '../actions';
-import inactiveFave from '../images/012-hollywood-star.svg';
-import activeFave from '../images/011-cinema.svg';
+import { loadFavorites } from '../../actions';
+import inactiveFave from '../../images/012-hollywood-star.svg';
+import activeFave from '../../images/011-cinema.svg';
 import {
   postFavorite,
   removeFavorite,
   fetchFavorites
-} from '../utilz/apiCalls';
+} from '../../utilz/apiCalls';
 
 class MoviePoster extends React.Component {
   constructor() {
@@ -42,10 +42,10 @@ class MoviePoster extends React.Component {
     const favoriteIDs = favorites.map(movie => movie.movie_id);
     if (!favoriteIDs.includes(movieID)) {
       this.addFavorite(userID);
-      this.setState({ isFavored: true})
+      this.setState({ isFavored: true });
     } else {
       this.deleteFavorite(userID, movieID);
-      this.setState({ isFavored: false})
+      this.setState({ isFavored: false });
     }
   }
 
@@ -66,9 +66,17 @@ class MoviePoster extends React.Component {
   render() {
     const { movie } = this.props;
     return (
-      <article className="movie-poster" >
-        <img src={this.state.isFavored ? activeFave : inactiveFave} className="favorite-icon" alt='fav-icon' onClick={e => this.handleFavorite(e)} />
-        <NavLink to={`/movies/${movie.movie_id}`} className='movie-showcase-link'>
+      <article className="movie-poster">
+        <img
+          src={this.state.isFavored ? activeFave : inactiveFave}
+          className="favorite-icon"
+          alt="fav-icon"
+          onClick={e => this.handleFavorite(e)}
+        />
+        <NavLink
+          to={`/movies/${movie.movie_id}`}
+          className="movie-showcase-link"
+        >
           <img
             src={`https://image.tmdb.org/t/p/w1280/${movie.poster_path}`}
             alt={`${movie.title}-poster`}
