+++ src/containers/App/App.js
@@ -1,15 +1,15 @@
 import React, { Component } from 'react';
-import { loadMovies, loadComplete, hasErrored } from '../actions';
-import { fetchMovieData } from '../utilz/apiCalls';
+import { loadMovies, loadComplete, hasErrored } from '../../actions';
+import { fetchMovieData } from '../../utilz/apiCalls';
 import { connect } from 'react-redux';
 import { Route } from 'react-router-dom';
-import loadingGif from '../images/loading.gif'
-import LoginForm from './LoginForm';
-import MoviesDisplay from './MoviesDisplay';
-import SignUpForm from './SignUpForm';
-import Favorites from '../components/Favorites';
-import Header from './Header';
-import MovieShowcase from '../components/MovieShowcase';
+import loadingGif from '../../images/loading.gif';
+import LoginForm from '../LoginForm/LoginForm';
+import MoviesDisplay from '../MoviesDisplay/MoviesDisplay';
+import SignUpForm from '../SignUpForm/SignUpForm';
+import Favorites from '../../components/Favorites/Favorites';
+import Header from '../../containers/Header/Header';
+import MovieShowcase from '../../components/MovieShowcase/MovieShowcase';
 
 class App extends Component {
   componentDidMount = () => {
@@ -28,28 +28,33 @@ class App extends Component {
   render() {
     const { isLoading, error } = this.props;
     const loading = (
-      <section className = 'loading-gif-container'>
+      <section className="loading-gif-container">
         <img src={loadingGif} alt="loading-gif" className="loading-gif" />
       </section>
-    )
+    );
     return (
       <main className="app">
         <Header />
         <h2 className="error">{error}</h2>
         <Route exact path="/login" render={() => <LoginForm />} />
         <Route exact path="/signup" render={() => <SignUpForm />} />
-        <Route exact path="/" render={() => !isLoading ? 
-          (<MoviesDisplay />) : 
-          loading}
+        <Route
+          exact
+          path="/"
+          render={() => (!isLoading ? <MoviesDisplay /> : loading)}
         />
-        <Route path='/movies/:id' render={({ match }) => {
-          const { id } = match.params;
-          const movie = this.props.movies.find(movie => movie.movie_id === parseInt(id));
-          console.log(movie)
-          return movie && <MovieShowcase movie={movie} />
-        }}
+        <Route
+          path="/movies/:id"
+          render={({ match }) => {
+            const { id } = match.params;
+            const movie = this.props.movies.find(
+              movie => movie.movie_id === parseInt(id)
+            );
+            console.log(movie);
+            return movie && <MovieShowcase movie={movie} />;
+          }}
         />
-        <Route exact path="/favorites" render={() => <Favorites /> } />
+        <Route exact path="/favorites" render={() => <Favorites />} />
       </main>
     );
   }
