+++ src/containers/LoginForm/LoginForm.js
@@ -1,21 +1,21 @@
 import React, { Component } from 'react';
+import { NavLink } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { postNewUser } from '../utilz/apiCalls';
-import { signIn, hasErrored, clearError } from '../actions';
-import { fetchUser } from '../utilz/apiCalls';
+import { fetchUser, fetchFavorites } from '../../utilz/apiCalls';
+import { signIn, hasErrored, clearError, loadFavorites } from '../../actions';
 import { withRouter } from 'react-router-dom';
 
-class SignUpForm extends Component {
+class LoginForm extends Component {
   constructor() {
     super();
     this.state = {
-      name: '',
       email: '',
       password: ''
     };
   }
 
   handleChange = e => {
+    e.preventDefault();
     this.setState({ [e.target.name]: e.target.value });
   };
 
@@ -34,36 +34,28 @@ class SignUpForm extends Component {
 
   handleSubmit = async e => {
     e.preventDefault();
-    const { name, email, password } = this.state;
-    const cleanEmail = this.handleEmail(email);
+    const { email, password } = this.state;
     try {
-      await postNewUser(name, cleanEmail, password);
-      let newUser = await fetchUser(cleanEmail, password);
-      this.props.signIn(newUser);
+      let user = await fetchUser(this.handleEmail(email), password);
+      this.props.signIn(user);
+      let favorites = await fetchFavorites(user.id);
+      this.props.loadFavorites(favorites);
       this.props.history.push('/');
       this.props.clearError();
     } catch ({ message }) {
       this.props.hasErrored(message);
-      this.props.history.push('/signup');
+      this.props.history.push('/login');
+      this.clearInputs();
     }
-    this.clearInputs();
   };
 
   clearInputs = () => {
-    this.setState({ name: '', email: '', password: '' });
+    this.setState({ email: '', password: '' });
   };
 
   render() {
     return (
       <form className="login-form">
-        <input
-          type="text"
-          name="name"
-          className="login-input"
-          placeholder="Name..."
-          onChange={this.handleChange}
-          value={this.state.name}
-        />
         <input
           type="email"
           name="email"
@@ -84,8 +76,13 @@ class SignUpForm extends Component {
           className="login-input login-btn"
           onClick={e => this.handleSubmit(e)}
         >
-          Create Your Account
+          Sign-In
         </button>
+        <NavLink to="/signup" className="login-create-acct-btn">
+          <p className="login-create-acct">
+            Don't have an account? Create one now!
+          </p>
+        </NavLink>
       </form>
     );
   }
@@ -99,12 +96,13 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch => ({
   signIn: user => dispatch(signIn(user)),
   hasErrored: errorMessage => dispatch(hasErrored(errorMessage)),
-  clearError: () => dispatch(clearError())
+  clearError: () => dispatch(clearError()),
+  loadFavorites: movies => dispatch(loadFavorites(movies))
 });
 
 export default withRouter(
   connect(
     mapStateToProps,
     mapDispatchToProps
-  )(SignUpForm)
+  )(LoginForm)
 );
