+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './containers/App';
+import App from './containers/App/App';
 import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
@@ -18,7 +18,7 @@ const application = (
       <App />
     </BrowserRouter>
   </Provider>
-)
+);
 
 ReactDOM.render(application, document.getElementById('root'));
